created PDF component and implemented into community and food page, centered pdf images
