+++ src/components/community.sass
@@ -28,6 +28,21 @@
   width: 100% 
   height: 230px
 
+.pdf-container
+  width: 100%
+  /* Firefox */
+  display: -moz-box
+  -moz-box-pack: center
+  -moz-box-align: center
+  /* Safari and Chrome */
+  display: -webkit-box
+  -webkit-box-pack: center
+  -webkit-box-align: center
+  /* W3C */
+  display: box
+  box-pack: center
+  box-align: center
+
 /* Responsive layout - makes a two column-layout instead of four columns */
 @media screen and (max-width: 800px)  
   .gallery-column
