+++ src/components/foodanddrink.sass
@@ -173,6 +173,22 @@ footer
   font-weight: 100 
   font-size: 50px 
   padding-top: 10% 
+  
+
+.pdf-container
+  width: 100%
+  /* Firefox */
+  display: -moz-box
+  -moz-box-pack: center
+  -moz-box-align: center
+  /* Safari and Chrome */
+  display: -webkit-box
+  -webkit-box-pack: center
+  -webkit-box-align: center
+  /* W3C */
+  display: box
+  box-pack: center
+  box-align: center
  
 
 @media only screen and (max-width: 768px)
